Split executor into writer executor and reader executor
