+++ crates/rooch-executor/src/actor/mod.rs
@@ -3,3 +3,4 @@
 
 pub mod executor;
 pub mod messages;
+pub mod reader_executor;
